Fix generator base-model list via MinIO bucket
