+++ backend/.env.example
@@ -14,3 +14,4 @@ MINIO_BUCKET_IMAGES=visionsuit-images
 MINIO_PUBLIC_URL=http://127.0.0.1:9000
 MINIO_AUTO_CREATE_BUCKETS=true
 # MINIO_REGION=us-east-1
+GENERATOR_BASE_MODEL_BUCKET=comfyui-models
